requirements.txt
asgiref==3.5.2
boto3==1.24.90
botocore==1.27.90
Django>=4.1.3,<4.2
django-admin-sortable2==2.1.4
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-storages==1.13.1
gunicorn>=20.1.0,<20.2
jmespath==1.0.1
Pillow==9.2.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
